+++ flutter/lib/desktop/screen/desktop_remote_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hbb/common.dart';
-import 'package:flutter_hbb/desktop/pages/connection_tab_page.dart';
+import 'package:flutter_hbb/desktop/pages/remote_tab_page.dart';
 import 'package:get/get.dart';
 import 'package:provider/provider.dart';
